Add warning about epan.h C #includes to README.md
When building Wireshark, the compiler throws an error for epan/epan.h.
The <glib.h> header, defined within the `extern "C"` block, has become
incompatible with pure C-style. To fix this, the `#include <glib.h>`
line had to be manually moved out of the `extern` block and recompiled.

Specs of the system on which the error was encountered:
* WSL: Ubuntu 22.04.2 LTS (GNU/Linux 5.15.133.1-microsoft-standard-WSL2
  x86_64)
* g++ (Ubuntu 11.4.0-1ubuntu1~22.04) 11.4.0
* gcc (Ubuntu 11.4.0-1ubuntu1~22.04) 11.4.0
* Wireshark version to install: 3.4.0

This warning should be removed if the change is fixed upstream (where
the Wireshark installation is sourced).